Refactor the cli layout
