+++ cli/.gitignore
@@ -1,2 +1,2 @@
 *.egg-info
-stage_variables/*.py
+stages/*.py
